requirements.txt
celery==5.1.2
Django==3.2.5
django-celery-beat==2.2.1
djangorestframework==3.12.4
drf-yasg==1.20.0
flower==1.0.0
redis==3.5.3
gunicorn==20.1.0
redisearch==2.0.0
colorama==0.4.4
termcolor==1.1.0
pyfiglet==0.8.post1
django-extensions==3.1.3
requests==2.26.0